api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.8.0
opik==1.0.0
supabase==2.3.0
Pillow>=10.4.0
python-multipart==0.0.6
